UI랑 일단 연결
