--- tests/ObjectDetection/Filtering.py
@@ -1,11 +1,0 @@
-import ObjectFiltering
-class Filtering:
-    
-    def __init__(self, savePath):
-        self.object = ObjectFiltering()
-        self.savePath = savePath
-        
-    def filtering(self, img, face, objects):
-        boxesList = []
-        boxesList.append(self.object.objectDetect(objects))
-        return boxesList
